--- policy/policy-02-provision/policy-provision-provider-policy-functions/build.gradle.kts
@@ -1,11 +1,0 @@
-plugins {
-    `java-library`
-    id("application")
-}
-
-dependencies {
-    api(libs.edc.data.plane.spi)
-
-    implementation(libs.edc.control.plane.core)
-
-}
