--- policy/policy-02-provision/policy-provision-provider-policy-functions/src/main/resources/META-INF/services/org.eclipse.edc.spi.system.ServiceExtension
@@ -1,1 +1,0 @@
-org.eclipse.sample.extension.provision.policy.PolicyFunctionsExtension
