+++ policy/policy-02-provision/policy-provision-provider/src/main/resources/META-INF/services/org.eclipse.edc.spi.system.ServiceExtension
@@ -1,0 +1,16 @@
+#
+#  Copyright (c) 2023 Fraunhofer Institute for Software and Systems Engineering
+#
+#  This program and the accompanying materials are made available under the
+#  terms of the Apache License, Version 2.0 which is available at
+#  https://www.apache.org/licenses/LICENSE-2.0
+#
+#  SPDX-License-Identifier: Apache-2.0
+#
+#  Contributors:
+#       Fraunhofer Institute for Software and Systems Engineering
+#
+#
+
+org.eclipse.edc.sample.extension.policy.PolicyFunctionsExtension
+org.eclipse.edc.sample.extension.transfer.FileTransferExtension
