--- policy/policy-02-provision/policy-transfer-file-local/src/main/java/org/eclipse/edc/sample/policy/file/extension/api/FileTransferDataSinkFactory.java
@@ -1,65 +1,0 @@
-/*
- *  Copyright (c) 2022 Microsoft Corporation
- *
- *  This program and the accompanying materials are made available under the
- *  terms of the Apache License, Version 2.0 which is available at
- *  https://www.apache.org/licenses/LICENSE-2.0
- *
- *  SPDX-License-Identifier: Apache-2.0
- *
- *  Contributors:
- *       Microsoft Corporation - initial API and implementation
- *
- */
-
-package org.eclipse.edc.sample.policy.file.extension.api;
-
-import org.eclipse.edc.connector.dataplane.spi.pipeline.DataSink;
-import org.eclipse.edc.connector.dataplane.spi.pipeline.DataSinkFactory;
-import org.eclipse.edc.spi.monitor.Monitor;
-import org.eclipse.edc.spi.result.Result;
-import org.eclipse.edc.spi.types.domain.transfer.DataFlowRequest;
-import org.jetbrains.annotations.NotNull;
-
-import java.io.File;
-import java.util.concurrent.ExecutorService;
-
-class FileTransferDataSinkFactory implements DataSinkFactory {
-    private final Monitor monitor;
-    private final ExecutorService executorService;
-    private final int partitionSize;
-
-    FileTransferDataSinkFactory(Monitor monitor, ExecutorService executorService, int partitionSize) {
-        this.monitor = monitor;
-        this.executorService = executorService;
-        this.partitionSize = partitionSize;
-    }
-
-    @Override
-    public boolean canHandle(DataFlowRequest request) {
-        return "file".equalsIgnoreCase(request.getDestinationDataAddress().getType());
-    }
-
-    @Override
-    public @NotNull Result<Void> validateRequest(DataFlowRequest request) {
-        return Result.success();
-    }
-
-    @Override
-    public DataSink createSink(DataFlowRequest request) {
-        var destination = request.getDestinationDataAddress();
-
-        // verify destination path
-        var path = destination.getStringProperty("path");
-        // As this is a controlled test input below is to avoid path-injection warning by CodeQL
-        var destinationFile = new File(path.replaceAll("\\.", ".").replaceAll("/", "/"));
-
-        return FileTransferDataSink.Builder.newInstance()
-                .file(destinationFile)
-                .requestId(request.getId())
-                .partitionSize(partitionSize)
-                .executorService(executorService)
-                .monitor(monitor)
-                .build();
-    }
-}
