--- policy/policy-02-provision/policy-transfer-file-local/src/main/resources/META-INF/services/org.eclipse.edc.spi.system.ServiceExtension
@@ -1,1 +1,0 @@
-org.eclipse.edc.sample.policy.file.extension.api.FileTransferExtension
