+++ settings.gradle.kts
@@ -56,8 +56,6 @@ include(":policy:policy-01-policy-enforcement:policy-functions")
 //policy
 
 include(":policy:policy-02-provision:policy-provision-consumer")
-include(":policy:policy-02-provision:policy-transfer-file-local")
-include(":policy:policy-02-provision:policy-provision-provider-policy-functions")
 include(":policy:policy-02-provision:policy-provision-provider")
 include(":policy:policy-02-provision:policy-provision-integration-tests")
